Initial Upload
If you want to run the script you only need to download Organize-Photos.ps1 and run it. 

1. Dump all your photos into a single folder. 
2. start the script.
3. Set options 1-4.
4. Run option 5! This will dump a file called test.csv into the same folder the script is being run from.
    Make sure you review that files, it will give you a better idea about whether you really should organize by creation date or modified 
    date.
5. Run the script with option 6. If you see an error message, figure it out on your own, its not too hard. I may be available to answer 
    questions, but don't count on that.

Gotachas
* You must set options 1-4 before you can run the script.
* Files in the destination folder will not be overwritten and a nasty error will appear, its ok, usually, you will need to manually review those 
   files.

TODO
*Logging
*Translation of filename to creation date

P.S.
I wrote this in an hour, and have not tested it in GNULinux/OSX, but it should run just fine so long as you use the modified attribute to organize files (creation time is not normally available). As I use Linux most of the time I will eventually assure that it does work in linux.